Comments Resolved
