+++ py-utils/src/utils/message_bus/message/__init__.py
@@ -1,5 +1,3 @@
 from __future__ import absolute_import
-from src.utils.message_bus.message.config import MessageConfig
-from src.utils.message_bus.message.header import MessageHeader
-from src.utils.message_bus.message.message import Message
-from src.utils.message_bus.message.format import MessageFormat
+
+from src.utils.message_bus.message.message import Message, MessageFormat, MessageConfig, MessageHeader
