--- py-utils/src/utils/message_bus/message/format.py
@@ -1,5 +1,0 @@
-
-class MessageFormat(object):
-    pass
-    # convert
-    #
